api/fastapi-reports/requirements.txt
fastapi
uvicorn
python-dotenv
python-keycloak
faker